+++ src/GiGraph.Dot.Entities/Attributes/Properties/Common/DotLabelAlignmentAttributes.cs
@@ -1,7 +1,6 @@
 using System.Reflection;
 using GiGraph.Dot.Entities.Attributes.Collections;
 using GiGraph.Dot.Entities.Attributes.Properties.KeyLookup;
-using GiGraph.Dot.Entities.Labels;
 using GiGraph.Dot.Output.Metadata;
 using GiGraph.Dot.Types.Alignment;
