+++ src/GiGraph.Dot.Output/IDotEscapable.cs
@@ -1,4 +1,4 @@
-﻿using GiGraph.Dot.Output.TextEscaping;
+﻿using GiGraph.Dot.Output.TextEscaping.Escapers;
 
 namespace GiGraph.Dot.Output
 {
