+++ src/GiGraph.Dot.Output/Options/DotSyntaxRules.Attributes.cs
@@ -1,4 +1,5 @@
-﻿using GiGraph.Dot.Output.TextEscaping;
+﻿using GiGraph.Dot.Output.TextEscaping.Escapers;
+using GiGraph.Dot.Output.TextEscaping.Pipelines;
 
 namespace GiGraph.Dot.Output.Options
 {
