--- src/GiGraph.Dot.Output/TextEscaping/DotQuotationMarkEscaper.cs
@@ -1,18 +1,0 @@
-﻿namespace GiGraph.Dot.Output.TextEscaping
-{
-    /// <summary>
-    ///     Escapes quotation marks.
-    /// </summary>
-    public class DotQuotationMarkEscaper : IDotTextEscaper
-    {
-        string IDotTextEscaper.Escape(string value)
-        {
-            return Escape(value);
-        }
-
-        public static string Escape(string value)
-        {
-            return value?.Replace("\"", "\\\"");
-        }
-    }
-}
