+++ src/GiGraph.Dot.Output/TextEscaping/Escapers/DotBackslashEscaper.cs
@@ -1,4 +1,4 @@
-﻿namespace GiGraph.Dot.Output.TextEscaping
+﻿namespace GiGraph.Dot.Output.TextEscaping.Escapers
 {
     /// <summary>
     ///     Escapes backslashes.
