+++ src/GiGraph.Dot.Output/TextEscaping/Escapers/DotCurlyBracketsEscaper.cs
@@ -1,4 +1,4 @@
-﻿namespace GiGraph.Dot.Output.TextEscaping
+﻿namespace GiGraph.Dot.Output.TextEscaping.Escapers
 {
     /// <summary>
     ///     Escapes curly brackets ({, }). Use for escaping text of record node fields.
