+++ src/GiGraph.Dot.Output/TextEscaping/Escapers/DotWindowsNewLineEscaper.cs
@@ -1,4 +1,4 @@
-﻿namespace GiGraph.Dot.Output.TextEscaping
+﻿namespace GiGraph.Dot.Output.TextEscaping.Escapers
 {
     /// <summary>
     ///     Escapes the carriage return + line feed character sequence (CRLF == \x000D\x000A == \r\n).
