+++ src/GiGraph.Dot.Output/TextEscaping/Escapers/Html/DotHtmlEncodeEscaper.cs
@@ -1,11 +1,11 @@
 ﻿using System.Web;
 
-namespace GiGraph.Dot.Output.TextEscaping
+namespace GiGraph.Dot.Output.TextEscaping.Escapers.Html
 {
     /// <summary>
     ///     Converts a string to an HTML-encoded string.
     /// </summary>
-    public class DotHtmlTextEscaper : IDotTextEscaper
+    public class DotHtmlEncodeEscaper : IDotTextEscaper
     {
         string IDotTextEscaper.Escape(string value)
         {
@@ -14,7 +14,7 @@ namespace GiGraph.Dot.Output.TextEscaping
 
         public static string Escape(string value)
         {
-            return value is not null ? HttpUtility.HtmlEncode(value) : value;
+            return value is not null ? HttpUtility.HtmlEncode(value) : null;
         }
     }
 }
