+++ src/GiGraph.Dot.Types/Text/DotConcatenatedEscapeString.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
-using GiGraph.Dot.Output.TextEscaping;
+using GiGraph.Dot.Output.TextEscaping.Escapers;
 
 namespace GiGraph.Dot.Types.Text
 {
